feat: Phase 0 scoring engine foundation
- Change CurrentStreak from int to float for weighted difficulty increments
- Apply DifficultyWeight to streak: MC=1.0, Cloze=1.2, Text=1.5, Sentence=2.5
- Implement temporal weighting: scaled wrong-answer penalty based on experience
- Implement partial streak preservation: experienced words keep some streak
- Add recovery-aware mastery: correct answers always show visible progress
- Fix VocabQuiz deferred recording write-back (currentItem.Progress = updatedProgress)
- Add EF Core migrations for SQLite and PostgreSQL (CurrentStreak INTEGER->REAL)
- Update tests for new scoring behavior (391/392 pass, 1 pre-existing failure)

+++ src/SentenceStudio.Contracts/Vocabulary/VocabularyProgressResponse.cs
@@ -9,6 +9,6 @@ public sealed class VocabularyProgressResponse
     public bool IsUserDeclared { get; set; }
     public DateTime? UserDeclaredAt { get; set; }
     public string VerificationState { get; set; } = string.Empty;
-    public int CurrentStreak { get; set; }
+    public float CurrentStreak { get; set; }
     public int TotalAttempts { get; set; }
 }
